+++ BackEnd/api/models/users.go
@@ -1,7 +1,7 @@
 package models
 
 type Users struct {
-	ID       int8   `json:"id"`
+	ID       int64  `json:"id"`
 	Name     string `json:"name"`
 	Email    string `json:"email"`
 	Password string `json:"password"`
